Added relevant game stats counters
Added code to keep track on the number of attempts the user has done to find card pairs, both positive and negative.
Also added a counter that keeps track of how many card pairs are left. This is also will work in the background to check for the win condition.
Both stats are visible to the user.